Moving Walk to Domain
