+++ pyprom/tests/testWalk.py
@@ -10,14 +10,12 @@ import unittest
 from pyprom.tests.getData import gettestzip
 from pyprom.dataload import GDALLoader
 from pyprom.feature_discovery import AnalyzeData
-from pyprom.walk import Walk
-from pyprom.lib.errors.errors import NoLinkersError
-from pyprom.lib.containers.runoffs import RunoffsContainer
+from pyprom.domain import Domain
 
 
 class WalkTests(unittest.TestCase):
     """
-    Test Walk Object.
+    Test The Walk Features of feature_discovery
     """
 
     def setUp(self):
@@ -31,29 +29,12 @@ class WalkTests(unittest.TestCase):
         self.islandpondVTVicinity = AnalyzeData(self.islandpondVT)
         self.summits, self.saddles, self.runoffs = \
             self.islandpondVTVicinity.run()
+        self.domain = Domain(self.islandpondVT, summits=self.summits,
+                             saddles=self.saddles, runoffs=self.runoffs,
+                             linkers=[])
+        self.no = None
 
-    def testWrongSaddles(self):
-        """
-        Ensure saddles param is actually a :class:SaddlesContainer
-        """
-        with self.assertRaises(TypeError):
-            Walk(self.summits, self.summits, self.runoffs, self.islandpondVT)
-
-    def testWrongSummits(self):
-        """
-        Ensure summits param is actually a :class:SummitsContainer
-        """
-        with self.assertRaises(TypeError):
-            Walk(self.saddles, self.saddles, self.runoffs, self.islandpondVT)
-
-    def testWrongDatamap(self):
-        """
-        Ensure datamap param is actually a :class:Datamap
-        """
-        with self.assertRaises(TypeError):
-            Walk(self.summits, self.saddles, self.runoffs, self.saddles)
-
-    def testWalkIslandPond(self):
+    def testWalkSingleSaddleIslandPond(self):
         """
         Test walk around Island Pond VT.
         """
@@ -61,40 +42,40 @@ class WalkTests(unittest.TestCase):
                                                         -71.8676388,
                                                         10)
         islandPondSaddle = islandPondSaddleContainer[0]
-        islandPondRunoff = RunoffsContainer([])
-        walk = Walk(self.summits, islandPondSaddleContainer,
-                    islandPondRunoff, self.islandpondVT)
-        walk.walk(islandPondSaddle)
-        self.assertEqual(len(walk.linkers), 2)
+        self.domain.walkSingleSaddle(islandPondSaddle)
+        self.assertEqual(len(self.domain.linkers), 2)
         self.assertEqual(len(islandPondSaddle.summits), 2)
-        self.assertEqual(islandPondSaddle.summits[0].summit, self.summits[155])
-        self.assertEqual(islandPondSaddle.summits[1].summit, self.summits[171])
+        self.assertEqual(islandPondSaddle.summits[0].summit,
+                         self.summits[155])
+        self.assertEqual(islandPondSaddle.summits[1].summit,
+                         self.summits[171])
 
-    def testGenerateDomainNoLinkers(self):
+    def testWalkIslandPond(self):
         """
-        Test that generating a domain from a Walk fails when the
-        walk has not been executed.
+        Test walk of island pond vicinity
         """
-        islandPondSaddleContainer = self.saddles.radius(44.810833,
-                                                        -71.8676388,
-                                                        10)
-        walk = Walk(self.summits, islandPondSaddleContainer,
-                    self.runoffs, self.islandpondVT)
-        with self.assertRaises(NoLinkersError):
-            walk.domain()
+        self.domain.walk()
+        self.assertEqual(len(self.domain.linkers), 1067)
 
-    def testGenerateDomain(self):
+    def testWalkMarkRedundantLinkers(self):
         """
-        Test that generating a domain from a properly executed Walk succeeds
+        Test mark_redundant_linkers()
         """
-        islandPondSaddleContainer = self.saddles.radius(44.810833,
-                                                        -71.8676388,
-                                                        10)
-        islandPondSaddle = islandPondSaddleContainer[0]
-        walk = Walk(self.summits, islandPondSaddleContainer,
-                    self.runoffs, self.islandpondVT)
-        walk.walk(islandPondSaddle)
-        d = walk.domain()
-        self.assertEqual(len(d.linkers), 2)
-        self.assertEqual(len(d.summits), 342)
-        self.assertEqual(len(d.saddles), 1)
+        self.domain.walk()
+        self.domain.mark_redundant_linkers()
+        self.assertEqual(len([x for x in self.domain.linkers
+                              if x.disqualified]), 0)
+        self.assertEqual(len([x for x in self.domain.linkers
+                              if x.disqualified]), 0)
+
+    def testWalkDisqualifyLowerLinkers(self):
+        """
+        Test disqualify_lower_linkers
+        """
+        self.domain.walk()
+        self.domain.disqualify_lower_linkers()
+        self.assertEqual(len([x for x in self.domain.linkers
+                              if x.disqualified]), 288)
+        self.domain.disqualify_lower_linkers()
+        self.assertEqual(len([x for x in self.domain.linkers
+                              if x.disqualified]), 288)
